api/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
bcrypt==4.2.1
gunicorn==23.0.0
